Tidy session file and create the idle timer
